Set a Dialog for the process wizard. And separate the code into different components
