+++ src/routes/+page.svelte
@@ -14,13 +14,10 @@
 	import { taiwanDayBoundsISO } from '$lib/utils/dates';
 	import { getMonthlyFromCacheFirst } from '$lib/data/monthly-cache-first';
 	import RetrieveExpenseButton from '$lib/components/RetrieveExpenseButton.svelte';
+	import AIReceiptImportDialog from '$lib/components/AIReceiptImportDialog.svelte';
 
-	import { PUBLIC_GOOGLE_AI_GCF, PUBLIC_SUPABASE_URL } from '$env/static/public';
-	import { supabase } from '$lib/supabase/supabaseClient';
 	import Logger from '$lib/utils/logger';
-	import * as jose from 'jose';
-	import { Sparkles, Upload, RotateCcw } from 'lucide-svelte';
-	import { browser } from '$app/environment';
+	import { Sparkles } from 'lucide-svelte';
 
 	let drawerOpen = $state(false);
 	let editMode = $state(false);
@@ -86,207 +83,16 @@
 	function openEdit(e: ExpenseRow) {
 		editMode = true;
 		expenseId = e.id;
-		// 初始化分帳
 		drawerOpen = true;
 	}
 
 	let aiDialogOpen = $state(false);
-	let aiUploading = $state(false);
-	let aiAnalyzing = $state(false);
-	let aiConverting = $state(false);
-	let selectedFile = $state<File | null>(null);
-	let previewUrl = $state<string | null>(null);
-	let isDragging = $state(false);
-	let fileInput: HTMLInputElement;
-	let lastUploadedFilePath = $state('');
-	// uploads/4b39017c-3ebb-466a-8ed2-3cbbfc8340e2/20260512032836_IMG_4146.jpg
 
-	interface ReceiptItem {
-		name: string;
-		quantity: number;
-		price: number;
-	}
-	interface ReceiptResult {
-		store_name: string;
-		date: string;
-		total_amount: number;
-		currency: string;
-		items: ReceiptItem[];
-	}
-	let analysisResult = $state<ReceiptResult | null>(null);
-
-	$effect(() => {
-		if (selectedFile) {
-			const url = URL.createObjectURL(selectedFile);
-			previewUrl = url;
-			return () => URL.revokeObjectURL(url);
-		} else {
-			previewUrl = null;
-		}
-	});
-
-	async function processFile(file: File) {
-		if (!browser) {
-			return;
-		}
-
-		const isHeic = file.name.toLowerCase().endsWith('.heic') || file.type === 'image/heic';
-		if (isHeic) {
-			aiConverting = true;
-			try {
-				// Dynamically import heic2any to avoid SSR issues
-				const heic2any = (await import('heic2any')).default;
-				const convertedBlob = await heic2any({
-					blob: file,
-					toType: 'image/jpeg',
-					quality: 0.7
-				});
-				const blob = Array.isArray(convertedBlob) ? convertedBlob[0] : convertedBlob;
-				selectedFile = new File([blob], file.name.replace(/\.heic$/i, '.jpg'), {
-					type: 'image/jpeg'
-				});
-				Logger.log('HEIC converted to JPEG successfully');
-			} catch (err) {
-				console.error('HEIC conversion failed:', err);
-				selectedFile = file;
-			} finally {
-				aiConverting = false;
-			}
-		} else {
-			selectedFile = file;
-		}
-	}
-
-	function handleFileChange(e: Event) {
-		const input = e.target as HTMLInputElement;
-		if (input.files && input.files[0]) {
-			processFile(input.files[0]);
-		}
-	}
-
-	function handleDrop(e: DragEvent) {
-		e.preventDefault();
-		isDragging = false;
-		if (e.dataTransfer?.files && e.dataTransfer.files[0]) {
-			processFile(e.dataTransfer.files[0]);
-		}
-	}
-
-	async function callAIGCF(fileName: string, contentType: string) {
-		const {
-			data: { session }
-		} = await supabase.auth.getSession();
-		const accessToken = session?.access_token;
-		if (!accessToken) {
-			console.error('No access token found');
-			return;
-		}
-
-		// verify if the access token is valid using jose and jwks
-		try {
-			const JWKS = jose.createRemoteJWKSet(
-				new URL(`${PUBLIC_SUPABASE_URL}/auth/v1/.well-known/jwks.json`)
-			);
-			const { payload } = await jose.jwtVerify(accessToken, JWKS);
-			Logger.log('Access token verified using jose:', payload);
-		} catch (error) {
-			console.error('Token verification failed:', error);
-			return;
-		}
-
-		try {
-			const url = new URL(PUBLIC_GOOGLE_AI_GCF);
-			url.searchParams.set('action', 'get_upload_url');
-			url.searchParams.set('file_name', fileName);
-			url.searchParams.set('content_type', contentType);
-
-			const response = await fetch(url.toString(), {
-				headers: {
-					Authorization: `Bearer ${accessToken}`
-				}
-			});
-			const signedUrl = await response.json();
-			Logger.log('Received signed URL from GCF:', signedUrl);
-			return signedUrl;
-		} catch (error) {
-			console.error('Error calling AI GCF:', error);
-			return null;
-		}
-	}
-
-	async function callAnalyzeReceipt(filePath: string) {
-		if (!filePath) {
-			return;
-		}
-		aiAnalyzing = true;
-		analysisResult = null;
-
-		const {
-			data: { session }
-		} = await supabase.auth.getSession();
-		const accessToken = session?.access_token;
-		if (!accessToken) {
-			console.error('No access token found');
-			aiAnalyzing = false;
-			return;
-		}
-
-		try {
-			const url = new URL(PUBLIC_GOOGLE_AI_GCF);
-			url.searchParams.set('action', 'analyze_receipt');
-			url.searchParams.set('file_path', filePath);
-
-			Logger.log('Triggering AI analysis for:', filePath);
-			const response = await fetch(url.toString(), {
-				headers: {
-					Authorization: `Bearer ${accessToken}`
-				}
-			});
-
-			const data = await response.json();
-			if (data.status === 'success' && data.result) {
-				analysisResult = data.result;
-			}
-			Logger.log('AI Analysis Result:', data);
-		} catch (error) {
-			console.error('Error calling analyze_receipt:', error);
-		} finally {
-			aiAnalyzing = false;
-		}
-	}
-
-	async function handleUpload() {
-		if (!selectedFile) {
-			return;
-		}
-		aiUploading = true;
-		try {
-			const contentType = selectedFile.type || 'image/jpeg';
-			const signedUrlData = await callAIGCF(selectedFile.name, contentType);
-
-			if (signedUrlData?.upload_url) {
-				const uploadRes = await fetch(signedUrlData.upload_url, {
-					method: 'PUT',
-					body: selectedFile,
-					headers: {
-						'Content-Type': contentType
-					}
-				});
-
-				if (uploadRes.ok) {
-					Logger.log('File uploaded to GCS successfully');
-					lastUploadedFilePath = signedUrlData.file_path;
-					// Trigger AI analysis automatically
-					await callAnalyzeReceipt(lastUploadedFilePath);
-				} else {
-					console.error('GCS Upload failed:', uploadRes.statusText);
-				}
-			}
-		} catch (err) {
-			console.error('Error in handleUpload:', err);
-		} finally {
-			aiUploading = false;
-		}
+	function handleImport(expenses: ExpenseRow[]) {
+		Logger.log('Importing expenses:', expenses);
+		// TODO: Implement actual database persistence for the imported items
+		// For now, we can just close the dialog. 
+		// You might want to add them to the store or trigger a refresh.
 	}
 
 	let startX = 0,
@@ -383,144 +189,7 @@
 	</Dialog.Content>
 </Dialog.Root>
 
-<Dialog.Root bind:open={aiDialogOpen}>
-	<Dialog.Content class="sm:max-w-[425px] {analysisResult ? 'max-h-[85vh] overflow-y-auto' : ''}">
-		<Dialog.Header>
-			<Dialog.Title>AI 記帳 - 上傳收據</Dialog.Title>
-		</Dialog.Header>
-
-		{#if analysisResult}
-			<div class="mt-4 space-y-4">
-				<div class="flex flex-col items-center">
-					<img
-						src={previewUrl}
-						alt="Receipt"
-						class="max-h-32 object-contain rounded shadow-sm"
-					/>
-				</div>
-				<div class="grid grid-cols-3 gap-2 text-sm border p-4 rounded-lg bg-muted/20">
-					<div class="text-muted-foreground text-xs uppercase">商店</div>
-					<div class="col-span-2 font-medium">{analysisResult.store_name}</div>
-					<div class="text-muted-foreground text-xs uppercase">日期</div>
-					<div class="col-span-2 font-medium">{analysisResult.date}</div>
-					<div class="text-muted-foreground text-xs uppercase">金額</div>
-					<div class="col-span-2 font-medium text-primary text-lg">
-						${analysisResult.total_amount}
-					</div>
-				</div>
-
-				{#if analysisResult.items && analysisResult.items.length > 0}
-					<div class="space-y-2">
-						<div class="text-xs font-semibold text-muted-foreground uppercase px-1">
-							項目明細
-						</div>
-						<div class="divide-y border rounded-lg overflow-hidden bg-muted/10">
-							{#each analysisResult.items as item, i (item.name + i)}
-								<div class="flex justify-between p-2 text-sm">
-									<div class="flex gap-2">
-										<span class="text-muted-foreground">x{item.quantity}</span>
-										<span>{item.name}</span>
-									</div>
-									<div class="font-medium">${item.price}</div>
-								</div>
-							{/each}
-						</div>
-					</div>
-				{/if}
-
-				<div class="flex gap-2 pt-2">
-					<Button
-						variant="outline"
-						class="grow"
-						onclick={() => (analysisResult = null)}
-						disabled={aiAnalyzing}
-					>
-						<RotateCcw class="w-4 h-4 mr-2" /> 重新分析
-					</Button>
-				</div>
-			</div>
-		{:else}
-			<div
-				class="mt-4 border-2 border-dashed rounded-lg p-4 flex flex-col items-center justify-center transition-colors {isDragging
-					? 'border-primary bg-primary/10'
-					: 'border-muted'}"
-				role="button"
-				tabindex="0"
-				onpointerenter={() => (isDragging = true)}
-				onpointerleave={() => (isDragging = false)}
-				ondragover={(e) => {
-					e.preventDefault();
-					isDragging = true;
-				}}
-				ondragleave={() => (isDragging = false)}
-				ondrop={handleDrop}
-				onclick={() => fileInput?.click()}
-				onkeydown={(e) => e.key === 'Enter' && fileInput?.click()}
-			>
-				<input
-					type="file"
-					accept="image/jpeg,image/png,image/heic,image/heif"
-					class="hidden"
-					bind:this={fileInput}
-					onchange={handleFileChange}
-				/>
-				{#if aiConverting}
-					<div class="flex flex-col items-center gap-2">
-						<div
-							class="h-8 w-8 animate-spin rounded-full border-4 border-primary border-t-transparent"
-						></div>
-						<p class="text-sm">轉換 HEIC 中...</p>
-					</div>
-				{:else if selectedFile && previewUrl}
-					<div class="flex flex-col items-center w-full">
-						<img
-							src={previewUrl}
-							alt="Preview"
-							class="max-h-48 w-full object-contain mb-2 rounded shadow-sm"
-						/>
-						<p class="text-sm font-medium truncate max-w-full">{selectedFile.name}</p>
-						<p class="text-xs text-muted-foreground">
-							{(selectedFile.size / 1024 / 1024).toFixed(2)} MB
-						</p>
-					</div>
-				{:else}
-					<Upload class="w-12 h-12 text-muted-foreground mb-2" />
-					<p class="text-sm">點擊或拖曳圖片至此</p>
-					<p class="text-xs text-muted-foreground mt-1">支援 JPG, PNG, HEIC 格式</p>
-				{/if}
-			</div>
-			<Dialog.Footer class="mt-4 flex flex-col gap-2">
-				<Button
-					class="w-full"
-					disabled={!selectedFile || aiUploading || aiConverting || aiAnalyzing}
-					onclick={handleUpload}
-				>
-					{#if aiUploading || aiAnalyzing}
-						<div class="flex items-center gap-2">
-							<div
-								class="h-4 w-4 animate-spin rounded-full border-2 border-current border-t-transparent"
-							></div>
-							{aiUploading ? '上傳中...' : '分析中...'}
-						</div>
-					{:else}
-						開始上傳並分析
-					{/if}
-				</Button>
-
-				{#if lastUploadedFilePath}
-					<Button
-						variant="secondary"
-						class="w-full"
-						disabled={aiAnalyzing}
-						onclick={() => callAnalyzeReceipt(lastUploadedFilePath)}
-					>
-						重新分析最後一次上傳
-					</Button>
-				{/if}
-			</Dialog.Footer>
-		{/if}
-	</Dialog.Content>
-</Dialog.Root>
+<AIReceiptImportDialog bind:open={aiDialogOpen} onImport={handleImport} />
 
 <style>
 </style>
